requirements.txt
python-dotenv>=0.10.5
praw>=6.5.1
